// nest-server/src/common/constants/actions.enum.ts
export enum ERoomActions {
  GET_MEMBER = 'GET_ROOM_MEMBER'
}

export enum ETaskTypeActions {
  READ = 'READ_TASK_TYPE',
  CREATE = 'CREATE_TASK_TYPE',
  UPDATE = 'UPDATE_TASK_TYPE',
  DELETE = 'DELETE_TASK_TYPE'
}

export enum ETaskActions {
  READ = 'READ_TASK',
  CREATE = 'CREATE_TASK',
  UPDATE = 'UPDATE_TASK',
  DELETE = 'DELETE_TASK'
}

export enum ESubTaskActions {
  UPDATE_STATUS = 'UPDATE_SUB_TASK_STATUS'
}

export enum EAwardActions {
  READ = 'READ_AWARD',
  RECEIVE = 'RECEIVE_AWARD',
  CREATE = 'CREATE_AWARD',
  UPDATE = 'UPDATE_AWARD',
  DELETE = 'DELETE_AWARD'
}

export enum ETaskCommentActions {
  READ = 'READ_COMMENT',
  CREATE = 'CREATE_COMMENT',
  UPDATE = 'UPDATE_COMMENT',
  DELETE = 'DELETE_COMMENT'
}